Getting rid of warnings pt2

+++ Source/Utilities/OverlappingSampleCollector.h
@@ -35,7 +35,7 @@ public:
     {
         jassert (overlap < bufferSize);
 
-        buffer.resize (bufferSize);
+        buffer.resize (uint(bufferSize));
         numCollected = 0;
 
         clearSkippedCounter();
@@ -88,11 +88,11 @@ public:
                     {
                         bufferQueue.push (buffer.data(), static_cast<int> (buffer.size())); // ... copy it into the queue
 
-                        const int overlapOffset = bufferSize - overlap; // where does the overlap start?
-                        for (int i = 0; i < overlap; ++i) // move the last overlap samples to the beginning
+                        const uint overlapOffset = uint(bufferSize - overlap); // where does the overlap start?
+                        for (uint i = 0; i < uint(overlap); ++i) // move the last overlap samples to the beginning
                             buffer[i] = buffer[overlapOffset + i];
 
-                        numCollected = overlap; // nice, we alreay have overlap many samples!
+                        numCollected = uint(overlap); // nice, we alreay have overlap many samples!
 
                         state = State::waitingForFreeSpace; // let's wait for new space
                         break;
